Move `home` component to `components/home` folder
