+++ apps/bestofjs-nextjs/src/components/home/home-featured-projects.client.tsx
